+++ app/models/token.rb
@@ -6,9 +6,6 @@ class Token < ApplicationRecord
   scope :active, -> { where(active: true).where('expires_at IS NULL OR expires_at >= ?', Time.current) }
   scope :expired, -> { where.not(expires_at: nil).where('expires_at < ?', Time.current) }
 
-  # relations
-  has_many :webhook_endpoints
-
   def self.expire!
     expired.update_all(active: false)
   end
