+++ db/migrate/20190429122649_create_webhook_endpoints.rb
@@ -3,7 +3,7 @@ class CreateWebhookEndpoints < ActiveRecord::Migration[5.2]
     create_table :webhook_endpoints, id: :uuid do |t|
       t.string :url, null: false
       t.string :events, null: false, array: true
-      t.references :token, index: true, foreign_key: true, type: :uuid, null: false
+      t.references :feed, index: true, foreign_key: true, type: :uuid, null: false
       t.timestamps
     end
